Changed unsigned longs and unsigned ints into size_t, whereever appropriate. Removed Reals and PlaceHolders from the aterm library. All tests on linux succeed. There are no non removable warnings on linux anymore.
Need feedback from the MSVC compiler regarding mismatches in 
the length of 32/64 integers.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8487 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/aterm/include/aterm/_aterm.h
@@ -12,7 +12,7 @@ typedef struct ProtEntry
 {
 	struct ProtEntry *next;
 	ATerm            *start;
-	unsigned long     size;
+	size_t     size;
 } ProtEntry;
 
 #define STRING_MARK 0xFF /* marker for binary strings */
@@ -22,10 +22,10 @@ extern ATbool silent;
 extern ATbool low_memory;
 
 extern ProtEntry **at_prot_table;
-extern unsigned int at_prot_table_size;
+extern size_t at_prot_table_size;
 extern ProtEntry *at_prot_memory;
 extern ATermProtFunc *at_prot_functions;
-extern unsigned int at_prot_functions_count;
+extern size_t at_prot_functions_count;
 
 void AT_markTerm(ATerm t);
 void AT_markTerm_young(ATerm t);
@@ -33,15 +33,15 @@ void AT_markTerm_young(ATerm t);
 void AT_unmarkTerm(ATerm t);
 void AT_unmarkIfAllMarked(ATerm t);
 void AT_unmarkAll();
-unsigned long  AT_calcTextSize(ATerm t);
-unsigned long  AT_calcCoreSize(ATerm t);
-unsigned long  AT_calcSubterms(ATerm t);
-unsigned long  AT_calcUniqueSubterms(ATerm t);
-unsigned long  AT_calcUniqueSymbols(ATerm t);
-unsigned long  AT_calcTermDepth(ATerm t);
+size_t  AT_calcTextSize(ATerm t);
+size_t  AT_calcCoreSize(ATerm t);
+size_t  AT_calcSubterms(ATerm t);
+size_t  AT_calcUniqueSubterms(ATerm t);
+size_t  AT_calcUniqueSymbols(ATerm t);
+size_t  AT_calcTermDepth(ATerm t);
 void AT_assertUnmarked(ATerm t);
 void AT_assertMarked(ATerm t);
-unsigned long AT_calcAllocatedSize();
+size_t AT_calcAllocatedSize();
 
 #ifdef __cplusplus
 }
